requirements.txt
streamlit
torch
torchvision
pandas
numpy
scikit-learn
matplotlib
plotly
Pillow
sqlite-utils
httpx>=0.27.0